+++ exchange/wrapper_bitfinex.go
@@ -251,6 +251,10 @@ func (w *bitfinexWrapper) GetFundingRate(ctx context.Context, symbol string) (fl
 	return w.adapter.GetFundingRate(ctx, symbol)
 }
 
+func (w *bitfinexWrapper) GetFundingInfo(ctx context.Context, symbol string) (*FundingInfo, error) {
+	return nil, ErrNotImplemented
+}
+
 func (w *bitfinexWrapper) GetIncomeHistory(ctx context.Context, symbol, incomeType string, startTime, endTime int64) ([]*income.Income, error) {
 	return nil, nil
 }