+++ exchange/wrapper_huobi.go
@@ -314,6 +314,10 @@ func (w *huobiWrapper) GetFundingRate(ctx context.Context, symbol string) (float
 	return w.adapter.GetFundingRate(ctx, symbol)
 }
 
+func (w *huobiWrapper) GetFundingInfo(ctx context.Context, symbol string) (*FundingInfo, error) {
+	return nil, ErrNotImplemented
+}
+
 func (w *huobiWrapper) GetIncomeHistory(ctx context.Context, symbol, incomeType string, startTime, endTime int64) ([]*income.Income, error) {
 	return nil, nil
 }