+++ exchange/wrapper_phemex.go
@@ -289,6 +289,10 @@ func (w *phemexWrapper) GetFundingRate(ctx context.Context, symbol string) (floa
 	return w.adapter.GetFundingRate(ctx)
 }
 
+func (w *phemexWrapper) GetFundingInfo(ctx context.Context, symbol string) (*FundingInfo, error) {
+	return nil, ErrNotImplemented
+}
+
 func (w *phemexWrapper) GetIncomeHistory(ctx context.Context, symbol, incomeType string, startTime, endTime int64) ([]*income.Income, error) {
 	return nil, nil
 }